requirements.txt
BeautifulSoup==3.2.1
Django==1.9.6
Markdown==2.6.6
Pillow==3.2.0
PyDispatcher==2.0.5
Scrapy==1.1.0
Twisted==16.2.0
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
boto==2.30.0
cffi==1.6.0
cryptography==1.3.2
cssselect==0.9.1
decorator==4.0.9
defusedxml==0.4.1
django-boto==0.3.1
django-filter==0.13.0
django-storages==1.1.8
django-suit==0.2.18
django-tastypie==0.13.3
djangorestframework==3.3.3
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
lxml==3.6.0
mock==1.0.1
parsel==1.0.2
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
python-dateutil==2.5.3
python-mimeparse==1.5.2
queuelib==1.4.2
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
w3lib==1.14.2
wsgiref==0.1.2
zope.interface==4.1.3
